rewrote apps handling to use floats. need to test

+++ nucleo-f767ZI-drev23-ecu-firmware/Core/Inc/app.h
@@ -19,7 +19,7 @@
 
 struct app_data {
 	// 
-	short torque;
+	float torque;
 	bool system_shutdown;
 	bool rtd_flag;
 	struct board board;
